+++ src/api/openapi.rs
@@ -1,29 +1,15 @@
 #![allow(non_snake_case)]
-use std::marker::PhantomData;
 
+#[cfg(feature = "openapi")]
+use std::marker::PhantomData;
+#[cfg(feature = "openapi")]
 use crate::{
-    api::{subresource::LoggingObject, Api, Object, RawApi, Void},
+    api::subresource::LoggingObject,
+    api::{Api, Object, Void},
     client::APIClient,
 };
+use crate::api::RawApi;
 use inflector::string::pluralize::to_plural;
-
-/// Implement a named constructor on Api for a typed kubernetes Object
-///
-/// This assumes that RawApi::$name exists first
-/// legacy, to be removed
-macro_rules! ctor {
-    ( $name:tt, $object:ty ) => {
-        impl Api<$object> {
-            pub fn $name(client: APIClient) -> Self {
-                Api {
-                    api: RawApi::$name(),
-                    client,
-                    phantom: PhantomData,
-                }
-            }
-        }
-    };
-}
 
 /// Bind typemeta properties for a k8s_openapi resource to RawApi
 ///
@@ -47,6 +33,7 @@ macro_rules! k8s_obj {
         }
     };
     // 3 argument version for empty prefix (lots of api::apps stuff has this)
+    // TODO: maybe apis should be default and  fix api::apps?
     ( $name:expr, $version:expr, $group:expr) => {
         k8s_obj!($name, $version, $group, "");
     };
@@ -58,6 +45,7 @@ macro_rules! k8s_obj {
 macro_rules! k8s_ctor {
     // using a standard openapi path with consistent Spec and Status suffixed structs
     ( $name:ident, $version:expr, $openapi:path) => {
+        #[cfg(feature = "openapi")]
         paste::item! {
             type [<Obj $name>] = Object<$openapi::[<$version>]::[<$name Spec>], $openapi::[<$version>]::[<$name Status>]>;
             impl Api<[<Obj $name>]> {
@@ -74,6 +62,7 @@ macro_rules! k8s_ctor {
 }
 
 /// Binds an arbitrary Object type to a verioned name on Api
+#[cfg_attr(not(feature = "openapi"), allow(unused_macros))]
 macro_rules! k8s_custom_ctor {
     // using a non-standard manual Object (for api inconsistencies)
     ( $versioned_name:ident, $obj:ty) => {
@@ -91,6 +80,7 @@ macro_rules! k8s_custom_ctor {
     };
 }
 
+
 // api::apps
 k8s_obj!("Deployment", "v1", "apps", "apis");
 k8s_ctor!(Deployment, "v1", k8s_openapi::api::apps);
@@ -103,21 +93,16 @@ k8s_ctor!(StatefulSet, "v1", k8s_openapi::api::apps);
 
 
 // api::authorization
-use k8s_openapi::api::authorization::v1 as v1Auth;
 k8s_obj!("SelfSubjectRulesReview", "v1", "authorization.k8s.io", "apis");
-k8s_custom_ctor!(v1SelfSubjectRulesReview, Object<v1Auth::SelfSubjectRulesReviewSpec, v1Auth::SubjectRulesReviewStatus>);
-#[test]
-fn k8s_obj_auth() {
-    let r = RawApi::v1SelfSubjectRulesReview();
-    assert_eq!(r.group, "authorization.k8s.io");
-    assert_eq!(r.prefix, "apis");
-    assert_eq!(r.resource, "selfsubjectrulesreviews"); // lowercase pluralisation
-}
-
+#[cfg(feature = "openapi")]
+k8s_custom_ctor!(v1SelfSubjectRulesReview, Object<k8s_openapi::api::authorization::v1::SelfSubjectRulesReviewSpec, k8s_openapi::api::authorization::v1::SubjectRulesReviewStatus>);
 
 // api::autoscaling
 k8s_obj!("HorizontalPodAutoscaler", "v1", "autoscaling", "apis");
 k8s_ctor!(HorizontalPodAutoscaler, "v1", k8s_openapi::api::autoscaling);
+
+// api::admissionregistration
+k8s_obj!("ValidatingWebhookConfiguration", "v1beta1", "admissionregistration.k8s.io", "apis"); // snowflake
 
 
 // api::core
@@ -133,16 +118,21 @@ k8s_obj!("PersistentVolume", "v1", "api");
 k8s_ctor!(PersistentVolume, "v1", k8s_openapi::api::core);
 k8s_obj!("ResourceQuota", "v1", "api");
 k8s_ctor!(ResourceQuota, "v1", k8s_openapi::api::core);
-
 k8s_obj!("PersistentVolumeClaim", "v1", "api");
 k8s_ctor!(PersistentVolumeClaim, "v1", k8s_openapi::api::core);
-
 k8s_obj!("ReplicationController", "v1", "api");
 k8s_ctor!(ReplicationController, "v1", k8s_openapi::api::core);
-
-use k8s_openapi::api::core::v1 as v1Core;
-impl LoggingObject for Object<v1Core::PodSpec, v1Core::PodStatus> {}
-impl LoggingObject for Object<v1Core::PodSpec, Void> {}
+// snowflakes in api::core
+k8s_obj!("Secret", "v1", "api");
+k8s_obj!("Event", "v1", "api");
+k8s_obj!("ConfigMap", "v1", "api");
+k8s_obj!("ServiceAccount", "v1", "api");
+k8s_obj!("Endpoints", "v1", "api"); // yup plural!
+// subresources
+#[cfg(feature = "openapi")]
+impl LoggingObject for Object<k8s_openapi::api::core::v1::PodSpec, k8s_openapi::api::core::v1::PodStatus> {}
+#[cfg(feature = "openapi")]
+impl LoggingObject for Object<k8s_openapi::api::core::v1::PodSpec, Void> {}
 
 
 // api::batch
@@ -156,14 +146,95 @@ k8s_ctor!(Job, "v1", k8s_openapi::api::batch);
 k8s_obj!("Ingress", "v1beta1", "extensions", "apis");
 k8s_ctor!(Ingress, "v1beta1", k8s_openapi::api::extensions);
 
+
 // apiextensions_apiserver::pkg::apis::apiextensions
 k8s_obj!("CustomResourceDefinition", "v1beta1", "apiextensions.k8s.io", "apis");
 k8s_ctor!(CustomResourceDefinition, "v1beta1", k8s_openapi::apiextensions_apiserver::pkg::apis::apiextensions);
 
+
+// api::rbac (snowflake objects in snowflake.rs)
+k8s_obj!("Role", "v1", "rbac.authorization.k8s.io", "apis");
+k8s_obj!("ClusterRole", "v1", "rbac.authorization.k8s.io", "apis");
+k8s_obj!("RoleBinding", "v1", "rbac.authorization.k8s.io", "apis");
+
+
 // api::storage::v1
-use k8s_openapi::api::storage::v1::{VolumeAttachmentSpec, VolumeAttachmentStatus};
-ctor!(v1VolumeAttachment, Object<VolumeAttachmentSpec, VolumeAttachmentStatus>);
+k8s_obj!("VolumeAttachment", "v1", "storage.k8s.io", "apis");
+k8s_ctor!(VolumeAttachment, "v1", k8s_openapi::api::storage);
+
 
 // api::networking::v1
-use k8s_openapi::api::networking::v1::NetworkPolicySpec;
-ctor!(v1NetworkPolicy, Object<NetworkPolicySpec, Void>); // has no Status
+k8s_obj!("NetworkPolicy", "v1", "networking.k8s.io", "apis");
+#[cfg(feature = "openapi")]
+k8s_custom_ctor!(v1NetworkPolicy, Object<k8s_openapi::api::networking::v1::NetworkPolicySpec, Void>); // no status
+
+
+// Macro insanity needs some sanity here..
+// There should be at least one test for each api group here to ensure no path typos
+mod test {
+    //use crate::api::{RawApi, PostParams};
+    // TODO: fixturize these tests
+    // these are sanity tests for macros that create the RawApi::v1Ctors
+    #[test]
+    fn api_url_secret() {
+        let r = RawApi::v1Secret().within("ns");
+        let req = r.create(&PostParams::default(), vec![]).unwrap();
+        assert_eq!(req.uri(), "/api/v1/namespaces/ns/secrets?");
+    }
+    #[test]
+    fn api_url_rs() {
+        let r = RawApi::v1ReplicaSet().within("ns");
+        let req = r.create(&PostParams::default(), vec![]).unwrap();
+        assert_eq!(req.uri(), "/apis/apps/v1/namespaces/ns/replicasets?");
+    }
+    #[test]
+    fn api_url_role() {
+        let r = RawApi::v1Role().within("ns");
+        let req = r.create(&PostParams::default(), vec![]).unwrap();
+        assert_eq!(req.uri(), "/apis/rbac.authorization.k8s.io/v1/namespaces/ns/roles?");
+    }
+    #[test]
+    fn api_url_cj() {
+        let r = RawApi::v1beta1CronJob().within("ns");
+        let req = r.create(&PostParams::default(), vec![]).unwrap();
+        assert_eq!(req.uri(), "/apis/batch/v1beta1/namespaces/ns/cronjobs?");
+    }
+    #[test]
+    fn api_url_hpa() {
+        let r = RawApi::v1HorizontalPodAutoscaler().within("ns");
+        let req = r.create(&PostParams::default(), vec![]).unwrap();
+        assert_eq!(req.uri(), "/apis/autoscaling/v1/namespaces/ns/horizontalpodautoscalers?");
+    }
+    #[test]
+    fn api_url_np() {
+        let r = RawApi::v1NetworkPolicy().within("ns");
+        let req = r.create(&PostParams::default(), vec![]).unwrap();
+        assert_eq!(req.uri(), "/apis/networking.k8s.io/v1/namespaces/ns/networkpolicies?");
+    }
+    #[test]
+    fn api_url_ingress() {
+        let r = RawApi::v1beta1Ingress().within("ns");
+        let req = r.create(&PostParams::default(), vec![]).unwrap();
+        assert_eq!(req.uri(), "/apis/extensions/v1beta1/namespaces/ns/ingresses?");
+    }
+    #[test]
+    fn api_url_vattach() {
+        let r = RawApi::v1VolumeAttachment();
+        let req = r.create(&PostParams::default(), vec![]).unwrap();
+        assert_eq!(req.uri(), "/apis/storage.k8s.io/v1/volumeattachments?");
+    }
+    #[test]
+    fn api_url_admission() {
+        let r = RawApi::v1beta1ValidatingWebhookConfiguration();
+        let req = r.create(&PostParams::default(), vec![]).unwrap();
+        assert_eq!(req.uri(), "/apis/admissionregistration.k8s.io/v1beta1/validatingwebhookconfigurations?");
+    }
+
+    #[test]
+    fn k8s_obj_custom_ctor() {
+        let r = RawApi::v1SelfSubjectRulesReview();
+        assert_eq!(r.group, "authorization.k8s.io");
+        assert_eq!(r.prefix, "apis");
+        assert_eq!(r.resource, "selfsubjectrulesreviews"); // lowercase pluralisation
+    }
+}