+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.26.0 / 2020-02-XX
 ===================
   * Exposed `Config::read` and `Config::read_from` - #124
+  * Fix typo on `Api::StatefulSet`
 
 0.25.0 / 2020-02-09
 ===================