+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Exposed `Config::read` and `Config::read_from` - #124
   * Fix typo on `Api::StatefulSet`
   * Fix typo on `Api::Endpoints`
+  * Add `Api::v1CustomResourceDefinition` when on k8s >= 1.17
 
 0.25.0 / 2020-02-09
 ===================